+++ src/obscure_stats/variation/variation.py
@@ -32,7 +32,7 @@ def mod_vr(x: np.ndarray) -> float:
     The Western Political Quarterly. 26 (2): 325-343.
     """
     cnts = np.unique(x, return_counts=True, equal_nan=True)[1]
-    return 1 - np.max(cnts) / len(x)
+    return float(1 - np.max(cnts) / len(x))
 
 
 def range_vr(x: np.ndarray) -> float:
@@ -61,7 +61,7 @@ def range_vr(x: np.ndarray) -> float:
     The Western Political Quarterly. 26 (2): 325-343.
     """
     cnts = np.unique(x, return_counts=True, equal_nan=True)[1]
-    return np.min(cnts) / np.max(cnts)
+    return float(np.min(cnts) / np.max(cnts))
 
 
 def gibbs_m1(x: np.ndarray) -> float:
@@ -103,7 +103,7 @@ def gibbs_m1(x: np.ndarray) -> float:
     Special case of Tsallis entropy (alpha = 2).
     """
     freq = np.unique(x, return_counts=True, equal_nan=True)[1] / len(x)
-    return 1 - np.sum(freq**2)
+    return float(1 - np.sum(freq**2))
 
 
 def gibbs_m2(x: np.ndarray) -> float:
@@ -133,7 +133,7 @@ def gibbs_m2(x: np.ndarray) -> float:
     """
     freq = np.unique(x, return_counts=True, equal_nan=True)[1] / len(x)
     k = len(freq)
-    return (k / (k - 1)) * (1 - np.sum(freq**2)) if k > 1 else 0
+    return float((k / (k - 1)) * (1 - np.sum(freq**2))) if k > 1 else 0.0
 
 
 def b_index(x: np.ndarray) -> float:
@@ -162,7 +162,7 @@ def b_index(x: np.ndarray) -> float:
     """
     n = len(x)
     freq = np.unique(x, return_counts=True, equal_nan=True)[1] / n
-    return 1 - (1 - (stats.gmean(freq * len(freq) / n)) ** 2) ** 0.5
+    return float(1 - (1 - (stats.gmean(freq * len(freq) / n)) ** 2) ** 0.5)
 
 
 def avdev(x: np.ndarray) -> float:
@@ -193,7 +193,7 @@ def avdev(x: np.ndarray) -> float:
     freq = np.unique(x, return_counts=True, equal_nan=True)[1] / n
     k = len(freq)
     mean = n / k
-    return 1 - (np.sum(np.abs(freq - mean)) / (2 * mean * max(k - 1, 1)))
+    return float(1 - (np.sum(np.abs(freq - mean)) / (2 * mean * max(k - 1, 1))))
 
 
 def renyi_entropy(x: np.ndarray, alpha: float = 2) -> float:
@@ -231,7 +231,7 @@ def renyi_entropy(x: np.ndarray, alpha: float = 2) -> float:
     if alpha == 1:
         # return Shannon entropy to avoid division by 0
         return -np.sum(freq * np.log2(freq))
-    return 1 / (1 - alpha) * math.log2(np.sum(freq**alpha))
+    return float(1 / (1 - alpha) * math.log2(np.sum(freq**alpha)))
 
 
 def negative_extropy(x: np.ndarray) -> float:
@@ -262,7 +262,7 @@ def negative_extropy(x: np.ndarray) -> float:
     """
     freq = np.unique(x, return_counts=True, equal_nan=True)[1] / len(x)
     p_inv = 1.0 - freq
-    return -np.sum(p_inv * np.log2(p_inv))
+    return float(-np.sum(p_inv * np.log2(p_inv)))
 
 
 def mcintosh_d(x: np.ndarray) -> float:
@@ -289,4 +289,4 @@ def mcintosh_d(x: np.ndarray) -> float:
     """
     n = len(x)
     counts = np.unique(x, return_counts=True, equal_nan=True)[1]
-    return (n - np.sum(counts**2) ** 0.5) / (n - n**0.5)
+    return float((n - np.sum(counts**2) ** 0.5) / (n - n**0.5))
